vi-rs への依存を削減
vi-rs で定義していた trait を interface に移動
vi-rs に依存するクレートを最小限に
