+++ gfx-interface/Cargo.toml
@@ -7,4 +7,3 @@ version = "0.1.0"
 winit = "0.26.0"
 bitflags ="1.3.2"
 raw-window-handle = "0.4.0"
-sjvi = { path = "../vi-rs" }
