+++ gfx-interface/src/swap_chain_api.rs
@@ -1,4 +1,4 @@
-use sjvi::IDisplayEventListener;
+use crate::IDisplayEventListener;
 
 pub struct SwapChainInfo {
     width: u32,
