+++ vi-rs/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2021"
 web-sys = { version = "0.3.58", features=[ "GpuBuffer", "HtmlCanvasElement", "WebGl2RenderingContext", "Window", "Document"] }
 wasm-bindgen = { version = "0.2" }
 winit = "0.26.1"
+sjgfx-interface = { path = "../gfx-interface" }
 
 [target.'cfg(not(any(target_arch = "wasm32")))'.dependencies]
 glutin = "*"
