+++ vi-rs/src/lib.rs
@@ -4,10 +4,7 @@
 pub mod winit;
 
 pub mod web_sys;
-
-pub trait IDisplayEventListener {
-    fn on_resized(&mut self, _width: u32, _height: u32) {}
-}
+use sjgfx_interface::IDisplay;
 
 pub trait IInstance {
     type DisplayId: Eq + PartialEq + Clone;
@@ -20,14 +17,6 @@ pub trait IInstance {
     fn try_get_display(&self, id: &Self::DisplayId) -> Option<&Self::Display>;
 
     fn try_update(&mut self) -> bool;
-}
-
-pub trait IDisplay {
-    fn is_redraw_requested(&self) -> bool;
-
-    fn listen<TListener: IDisplayEventListener>(&self, listener: &mut TListener);
-
-    fn get_scale_factor(&self) -> f64;
 }
 
 #[cfg(test)]
